+++ CHANGELOG.md
@@ -39,6 +39,7 @@ Improvements
 - Use max and min MaxRecursionDepth rather than sum and difference when combining UiMetrics
 - Avoid unnecessary enumerator allocations when updating and drawing elements
 - Allow setting a custom link color per paragraph
+- Made Element TransformInverse and TransformInverseAll public
 
 Fixes
 - Fixed tooltips not being bounded correctly for viewports that don't start at the origin